First feature of acronym bot
When a user types the slash command `/whatis` followed by an acronym, the bot
will display the expansion of that acronym. If the acronym is not present on
the yaml file, the bot will display a message to inform the user.

This is the first user journey/basic feature of the acronym bot.