Bot sends messages in-channel
Now when a user requests the expansion of an acronym, the Acronym bot will reply
in the channel rather than sending an ephemeral response back to the user.
To keep the requests anonymous, only the bot response will be display in the
channel, omiting the slash command that triggered that response.