Install puma
All the tests passed, deploying to Heroku worked, and yet the app
crashed because it `Couldn't find handler for: puma, thin, falcon,
webrick.`

Trying to run it locally revealed the same error. Not sure why it used
to work - perhaps Heroku used to have a default app server installed on
the Ruby stack?